requirements.txt
asgiref==3.7.2
certifi==2025.8.3
cloudinary==1.41.0
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.2.0
django-crispy-forms==2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
itsdangerous==2.2.0
packaging==25.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2025.2
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.6.0
